// frontend/src/gen/types/RAGCollectionListSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { RAGDocumentListSchema } from "./RAGDocumentListSchema.ts";

export type RAGCollectionListSchema = {
    /**
     * @type array | undefined
    */
    documents?: RAGDocumentListSchema[];
    id?: (number | null);
    /**
     * @maxLength 255
     * @type string
    */
    rag_collection_name: string;
    /**
     * @maxLength 255
     * @default ""
     * @type string | undefined
    */
    qdrant_collection_name?: string;
};