// frontend/src/gen/types/IndexingStatusResponseSchema.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const indexingStatusResponseSchemaStatusEnum = {
    "PENDING": "PENDING",
    "PROGRESS": "PROGRESS",
    "SUCCESS": "SUCCESS",
    "FAILURE": "FAILURE"
} as const;

export type IndexingStatusResponseSchemaStatusEnumKey = (typeof indexingStatusResponseSchemaStatusEnum)[keyof typeof indexingStatusResponseSchemaStatusEnum];

export type IndexingStatusResponseSchema = {
    /**
     * @type string
    */
    status: IndexingStatusResponseSchemaStatusEnumKey;
    progress?: (number | null);
    message?: (string | null);
};